Most of the effects system has been implemented.

+++ src/main/kotlin/com/github/ryuzu/ryuzutechnicalmagiccore/core/model/entity/effect/StatusEffectState.kt
@@ -2,7 +2,7 @@ package com.github.ryuzu.ryuzutechnicalmagiccore.core.model.entity.effect
 
 data class StatusEffectState(
     val id: StatusEffectId,
-    var duration: Long,
+    var duration: Int,
     var level: Int,
     val startedAt: Long = System.currentTimeMillis(),
 ) {
